Move DimensionalMetricAggregator to merics package
