+++ agent-bridge/src/main/java/com/newrelic/agent/bridge/NoOpAgent.java
@@ -8,7 +8,7 @@
 package com.newrelic.agent.bridge;
 
 import com.newrelic.api.agent.Config;
-import com.newrelic.api.agent.DimensionalMetricAggregator;
+import com.newrelic.api.agent.metrics.DimensionalMetricAggregator;
 import com.newrelic.api.agent.ErrorApi;
 import com.newrelic.api.agent.Insights;
 import com.newrelic.api.agent.Logger;
