+++ newrelic-api/src/main/java/com/newrelic/api/agent/metrics/DimensionalMetricAggregator.java
@@ -1,4 +1,4 @@
-package com.newrelic.api.agent;
+package com.newrelic.api.agent.metrics;
 
 import java.util.Map;
